result buttons

+++ ddgcalc.py
@@ -5,14 +5,16 @@ import os
 import requests as r
 import subprocess
 import sys
+import os.path
+from os import path
 
 aminoacids = {'Ala': 'A', 'Arg': 'R', 'Asn': 'N', 'Asp': 'D', 'Asx': 'B', 'Cys': 'C', 'Glu': 'E', 'Gln': 'Q', 'Glx': 'Z', 'Gly': 'G', 'His': 'H',
               'Ile': 'I', 'Leu': 'L', 'Lys': 'K', 'Met': 'M', 'Phe': 'F', 'Pro': 'P', 'Ser': 'S', 'Thr': 'T', 'Trp': 'W', 'Tyr': 'Y', 'Val': 'V', }
 
-def ddgcalcs(pdb, aasub, gene):
+def ddgcalcs(pdb, aasub, gene, use_mut):
     # parse amino acid substitution -> wild_position_mutant
-    wild = aminoacids.get(aasub[:3])
-    mutant = aminoacids.get(aasub[-3:])
+    wild = aminoacids.get(aasub[:3]) if use_mut == True else aminoacids.get(aasub[-3:])
+    mutant = aminoacids.get(aasub[-3:]) if use_mut == True else aminoacids.get(aasub[:3])
     position = int(aasub[3:-3])
 
     # retrieve pdb file
@@ -46,6 +48,7 @@ def ddgcalcs(pdb, aasub, gene):
         elif len(mol) < 5 and (gene in mol[2] or uniprotcode in mol[1]):
             shift = int(mol[3]) - int(mol[0])
             uniprotcode = mol[1]
+        
             
     # SAAMBE-3D calculation
     saambe_out = subprocess_cmd('source /Users/cameosameshima/opt/anaconda3/etc/profile.d/conda.sh; conda activate py2;\
@@ -60,7 +63,8 @@ def ddgcalcs(pdb, aasub, gene):
     #     saambe_eff = "Decrease in stability" if float(saambe_val) > 0.0 else "Increase in stability"
     # else:
     #     saambe_val = saambe_eff = 'N/A'
- 
+    print(saambe_out)
+    
     # imut2.0 struc calculation
     subprocess_cmd('source /Users/cameosameshima/opt/anaconda3/etc/profile.d/conda.sh; conda activate py2;\
         cd prediction/imutant;\
@@ -80,6 +84,7 @@ def ddgcalcs(pdb, aasub, gene):
     #     imut2_val = imut2_eff = 'N/A'
     imut2_val = (imut2_out.decode("utf-8").split("RSA")[1].split("WT")[0]).split()[3] if "I-Mutant" in imut2_out.decode("utf-8") else 'N/A'
     imut2_eff = muteffect(imut2_val,False) if imut2_val != 'N/A' and muteffect(imut2_val,False) else 'N/A'
+    print(imut2_out)
 
     # imut2.0 seq calculation
     sequence = getsequence(uniprotcode)
@@ -101,41 +106,58 @@ def ddgcalcs(pdb, aasub, gene):
         imut2_seq_val ='N/A'
     # imut2_seq_val = (imut2_seq_out.decode("utf-8").split("RSA")[1].split("WT")[0]).split()[3] if "I-Mutant" in imut2_seq_out.decode("utf-8") else 'N/A'
     imut2_seq_eff = muteffect(imut2_seq_val,False) if  imut2_seq_val != 'N/A' and muteffect(imut2_seq_val,False) else 'N/A'
+    print(imut2_seq_val)
 
     # UEP calculation
     os.system('cd prediction/uep; python3 UEP.py --pdb=../tmp/'+pdb +'.pdb --interface='+chain+','+secondchain)
-    uep_out = pd.read_csv('prediction/tmp/'+pdb+'_UEP_'+chain+'_'+secondchain+'.csv')
-    location = uep_out.loc[uep_out['Unnamed: 0'] == chain+'_' + str(position - shift)+'_'+aasub[:3].upper(), aasub[-3:].upper()]
-    uep_val = "N/A" if location.empty else location.values[0]
+    uep_file_path = 'prediction/tmp/'+pdb+'_UEP_'+chain+'_'+secondchain+'.csv'
+    if path.exists(uep_file_path):
+        uep_out = pd.read_csv(uep_file_path)
+        location = uep_out.loc[uep_out['Unnamed: 0'] == chain+'_' + str(position - shift)+'_'+aasub[:3].upper(), aasub[-3:].upper()]
+        uep_val = "N/A" if location.empty else location.values[0]
+        uep_eff = muteffect(uep_val,True) if uep_val != "N/A" and muteffect(uep_val,True) else 'N/A'
+        os.remove('prediction/tmp/'+pdb+'_UEP_'+chain+'_'+secondchain+'.csv')
+    else:
+        uep_val = uep_eff = 'N/A'
+
     # if uep_val != "N/A":
     #     uep_eff = "Improved binding affinity" if float(uep_val) < 0 else "Improved binding affinity" 
     # else:
     #     uep_eff = "N/A"
-    uep_eff = muteffect(uep_val,True) if uep_val != "N/A" and muteffect(uep_val,True) else 'N/A'
-
+    
     # panda calculation
-    secondseq = "'" + getsequence(secounduniprotcode) + "'"
+    
     mutseq = "'" + sequence[:position-1] + mutant + sequence[position:] + "'"
-    sequence = "'" + sequence + "'"
+    sequence = "'" + sequence[:position-1] + wild + sequence[position:] + "'"
+    secondseq = "'" + getsequence(secounduniprotcode) + "'"
 
     panda_output = subprocess_cmd('cd prediction/panda;\
         (echo "from panda import *"; echo "print(predict_affinity('+secondseq+','+sequence+','+secondseq+','+mutseq+'))") | python')
-    panda_val = panda_output.decode("utf-8")[1:-1]
+    panda_val = panda_output.decode("utf-8") 
+    if panda_val != "0":
+        panda_val = panda_output.decode("utf-8")[1:-1]
+
     # if panda_val:
     #     panda_eff = "Increase in affinity" if float(panda_val) > 0.0 else "Decrease in affinity"
     # else:
     #     panda_val = panda_eff = 'N/A'
+    print(panda_output)
+    print(panda_val)
     panda_eff = muteffect(panda_val,False) if panda_val != 'N/A' and muteffect(panda_val,False) else 'N/A'
 
     # Consensus
     increase = len(list(filter(lambda x: x == "Increase in stability", [saambe_eff, imut2_eff, imut2_seq_eff, panda_eff, uep_eff])))
     decrease = len(list(filter(lambda x: x == "Decrease in stability", [saambe_eff, imut2_eff, imut2_seq_eff, panda_eff, uep_eff])))
-    consensus = "Increase in stability" if increase > decrease else "Decrease in stability"
+    if increase > decrease:
+        consensus = "Increase in stability"
+    elif decrease > increase:
+        consensus = "Decrease in stability"
+    else:
+        consensus = "No change in stability"
 
     # get rid of files
     os.remove('prediction/tmp/'+pdb+'.pdb')
     os.remove('prediction/tmp/'+pdb+'.dssp')
     os.remove('prediction/tmp/sequence.seq')
-    os.remove('prediction/tmp/'+pdb+'_UEP_'+chain+'_'+secondchain+'.csv')
 
     return [["SAAMBE-3D",saambe_val, saambe_eff], ["I-Mutant2.0 Structure", imut2_val, imut2_eff], ["I-Mutant2.0 Sequence",imut2_seq_val, imut2_seq_eff], ["PANDA", panda_val, panda_eff], ["UEP",uep_val, uep_eff],["Stability Consensus","",consensus]], chain
